Finished backend(hopefully); Added likes/commentaries

+++ server/controllers/PostController.js
@@ -1,4 +1,5 @@
 import PostModel from '../models/Post.js';
+import CommentModel from '../models/Comment.js';
 
 export const getLastTags = async (req, res) => {
   try {
@@ -20,12 +21,14 @@ export const getLastTags = async (req, res) => {
 
 export const getAll = async (req, res) => {
   try {
-    const posts = await PostModel.find().populate('user').exec();
+    const posts = await PostModel.find()
+      .populate('user').populate('comments')
+      .exec();
     res.json(posts);
   } catch (err) {
     console.log(err);
     res.status(500).json({
-      message: 'Не удалось получить статьи',
+      message: 'Не удалось получить данные',
     });
   }
 };
@@ -34,15 +37,31 @@ export const getOne = async (req, res) => {
   try {
     const postId = req.params.id;
 
-    PostModel.findOneAndUpdate(
+    await PostModel.findOneAndUpdate(
       { _id: postId }, { $inc: { viewsCount: 1 } }, { returnDocument: "After" })
+      .populate('user') // only works for model type, then will change the type to json
+      .populate('comments')
       .then(doc => res.json(doc))
-      .catch(err => res.status(500).json({ message: "Не удалось получить статьи" }))
-      .populate('user');
+      .catch(err => res.status(500).json({ message: "Не удалось получить данные" }));
   } catch (err) {
     console.log(err);
     res.status(500).json({
-      message: 'Не удалось получить статьи',
+      message: 'Не удалось получить данные',
+    });
+  }
+};
+
+export const getByTags = async (req, res) => {
+  try {
+    const tags = req.query.tags;
+    await PostModel.find({ tags: { "$in": tags } }).then(doc => res.json(doc));
+    console.log(tags);
+    //return res.json(tags);
+  }
+  catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: 'Не удалось получить данные',
     });
   }
 };
@@ -51,13 +70,26 @@ export const remove = async (req, res) => {
   try {
     const postId = req.params.id;
 
+    let checkUser = await PostModel.findOne({ _id: [postId] }).then(doc => {
+      if (doc !== null) {
+        return doc.user._id.toString();
+      }
+      else {
+        throw new Error("Пост не найден");
+      }
+    });
+
+    if ((checkUser !== req.userId)) {
+      return res.status(500).json({ message: "Нет доступа" });
+    }
+
     PostModel.findOneAndDelete({ _id: postId, })
-      .then(doc => res.status(200).json({ message: "success!" }))
-      .catch(err => res.status(500).json({ message: "Не удалось удалить статью" }));
+      .then(doc => res.status(200).json({ message: "Пост успешно удален!" }))
+      .catch(err => res.status(500).json({ message: "Не удалось удалить данные" }));
   } catch (err) {
     console.log(err);
     res.status(500).json({
-      message: 'Не удалось получить статьи',
+      message: 'Не удалось получить данные',
     });
   }
 };
@@ -78,7 +110,7 @@ export const create = async (req, res) => {
   } catch (err) {
     console.log(err);
     res.status(500).json({
-      message: 'Не удалось создать статью',
+      message: 'Не удалось создать данные',
     });
   }
 };
@@ -86,6 +118,20 @@ export const create = async (req, res) => {
 export const update = async (req, res) => {
   try {
     const postId = req.params.id;
+    let checkUser = await PostModel.findOne({ _id: [postId] }).
+      then(doc => {
+        if (doc !== null) {
+          return doc.user._id.toString();
+        }
+        else {
+          throw new Error("Пост не найден");
+        }
+      });
+
+    console.log(checkUser + " " + req.userId + "| RES="+ (checkUser===req.userId));
+    if ((checkUser !== req.userId)) {
+      return res.status(500).json({ message: "Нет доступа" });
+    }
 
     await PostModel.updateOne(
       {
@@ -98,15 +144,178 @@ export const update = async (req, res) => {
         user: req.userId,
         tags: req.body.tags.split(','),
       },
-    );
+    ).then(doc => res.json(doc));
 
-    res.json({
-      success: true,
-    });
   } catch (err) {
     console.log(err);
     res.status(500).json({
-      message: 'Не удалось обновить статью',
+      message: 'Не удалось обновить данные',
+    });
+  }
+};
+
+
+export const likePost = async (req, res) => {
+  try {
+    const postId = req.params.id;
+
+    let post = await PostModel.findOne({ _id: postId }).
+      then(doc => {
+        if (doc !== null) {
+          return doc;
+        }
+        else {
+          throw new Error("Пост не найден");
+        }
+      });
+    if (post.likesCount.indexOf(req.userId) === -1) {
+      post.likesCount.push(req.userId);
+    }
+    else {
+      post.likesCount.splice(post.likesCount.indexOf(req.userId), 1);
+    }
+
+    const data = await post.save();
+
+    console.log(post.likesCount)
+
+    return res.json(data);
+
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: 'Не удалось обновить данные',
+    });
+  }
+};
+
+export const createComment = async (req, res) => {
+  try {
+    const postId = req.params.id;
+
+    const doc = new CommentModel({
+      text: req.body.text,
+      user: req.userId,
+    });
+
+    const comment = await doc.save();
+    await PostModel.updateOne({ _id: postId }, { $push: { comments: doc } })
+      .then(doc => console.log("Added comment to POST #" + postId))
+
+    res.json(comment);
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: 'Не удалось создать комментарий',
+    });
+  }
+};
+
+export const getAllComments = async (req, res) => {
+  try {
+    await CommentModel.find()
+      .populate('user') // only works for model type, then will change the type to json
+      .then(doc => res.json(doc));
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: 'Не удалось получить данные',
+    });
+  }
+};
+
+export const updateComment = async (req, res) => {
+  try {
+    const commentId = req.params.id;
+    let checkUser = await CommentModel.findOne({ _id: [commentId] }).
+      then(doc => {
+        if (doc !== null) {
+          doc.user._id.toString();
+        }
+        else {
+          throw new Error("Комментарий не найден");
+        }
+      });
+
+    if ((checkUser !== req.userId)) {
+      return res.status(500).json({ message: "Нет доступа" });
+    }
+
+    await CommentModel.updateOne(
+      {
+        _id: commentId,
+      },
+      {
+        text: req.body.text
+      },
+    ).then(doc => res.json(doc));
+
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: 'Не удалось обновить данные',
+    });
+  }
+};
+
+export const likeComment = async (req, res) => {
+  try {
+    const commentId = req.params.id;
+
+    let comment = await CommentModel.findOne({ _id: commentId }).
+      then(doc => {
+        if (doc !== null) {
+          return doc;
+        }
+        else {
+          throw new Error("Комментарий не найден");
+        }
+      });
+    if (comment.likesCount.indexOf(req.userId) === -1) {
+      comment.likesCount.push(req.userId);
+    }
+    else {
+      comment.likesCount.splice(comment.likesCount.indexOf(req.userId), 1);
+    }
+
+    const data = await comment.save();
+
+    console.log(comment.likesCount)
+
+    return res.json(data);
+
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: 'Не удалось обновить данные',
+    });
+  }
+};
+
+export const removeComment = async (req, res) => {
+  try {
+    const commentId = req.params.id;
+
+    let checkUser = await CommentModel.findOne({ _id: [commentId] }).then(doc => {
+      if (doc !== null) {
+        doc.user._id.toString();
+      }
+      else {
+        throw new Error("Комментарий не найден");
+      }
+    });
+
+    if ((checkUser !== req.userId)) {
+      return res.status(500).json({ message: "Нет доступа" });
+    }
+
+    CommentModel.findOneAndDelete({ _id: commentId, })
+      .then(doc => res.status(200).json({ message: "Комментарий успешно удален!" }))
+      .catch(err => res.status(500).json({ message: "Не удалось удалить комментарий" }));
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: 'Не удалось получить данные',
     });
   }
 };
